Separated API views from the ones rendering other views. API paths are changed to include 'api/' in them.
