+++ feat/api_urls.py
@@ -1,10 +1,10 @@
 from django.urls import include, path
 from rest_framework import routers
-from . import views
+from . import api_views
 
 router = routers.DefaultRouter()
-router.register(r'consumer_profile', views.ConsumerProfileViewSet)
-router.register(r'provider_profile', views.ProviderProfileViewSet)
+router.register(r'consumer_profile', api_views.ConsumerProfileViewSet)
+router.register(r'provider_profile', api_views.ProviderProfileViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
